Re-upload with GitKracken
This is a few days in, I was told to upload things to GitHub each day I work on them. So this is the start of that. 
I've got doc info inside the code, but short version is this site I'm working on will have basically a book in it that you can pull from for personal reasons. It's for a role-playing World of Darkness system. Eventually I'll have all of the SQL database, code, and html up and running so that users can search for content, navigate to it via a nav bar, hit a button to copy, edit, and save text, have a random dice roller for the system, and a character sheet as well. It's a lofty project, and I'm pretty much teaching myself as I go along.