Index update, CSS files, and new code
Revision to the index.html to make it more visible to what it was.
Broke out CSS into a separate file and created a second CSS sheet for printing from the mini-notepad area.
Added the noteArea. A textArea that has three buttons to allow the user to open a previous note, save a note (in JSON format), and a print button that will grab all the text from the textArea, then open a new window giving just the header and text there. Printing from the page as it was, there were too many dark inks, and wanted something cleaner looking. Still learning on how to save the files, that will be my next task. 
Finally added calls to get to the noteArea and back to main article (since I'm reusing the same space). As it stands there's only one button to get into the notes, (I might add another one under the chapter accordions) but on the note pane, there's the button at the nav bar and any of the links on the left will port the user back to the articles.